First set of changes to bring unit test naming up to the standards in the style guide
