+++ lib/improver/tests/convection/test_DiagnoseConvectivePrecipitation.py
