+++ lib/improver/tests/generate_ancillaries/test_generate_ancillary.py
@@ -1,4 +1,4 @@
-# -*- coding: utf-8 -*-
+# -*- coding: iso-8859-1 -*-
 # -----------------------------------------------------------------------------
 # (C) British Crown Copyright 2017 Met Office.
 # All rights reserved.
@@ -28,8 +28,7 @@
 # CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
 # ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
-"""Unit tests for the generate_ancillary.find_standard_ancil function."""
-
+"""Unit tests for the the stand alone functions in generate_ancillary.py"""
 
 import unittest
 from glob import glob
@@ -43,8 +42,7 @@ from iris import save
 import numpy as np
 
 from improver.generate_ancillaries.generate_ancillary import (
-    find_standard_ancil)
-
+    _make_mask_cube, find_standard_ancil)
 
 def _make_test_cube(long_name, stash=None):
     cs = GeogCS(EARTH_RADIUS)
@@ -61,6 +59,50 @@ def _make_test_cube(long_name, stash=None):
     if stash is not None:
         cube.attributes['STASH'] = stash
     return cube
+
+
+class Test__make_mask_cube(IrisTest):
+    """Test private function to make cube from generated mask"""
+
+    def setUp(self):
+        """setting up test data"""
+        premask = np.array([[0., 3., 2.],
+                            [0.5, 0., 1.5],
+                            [0.2, 0., 0]])
+        self.mask = np.ma.masked_where(premask > 1., premask)
+        self.key = 'test key'
+        self.x_coord = DimCoord([1, 2, 3], long_name='longitude')
+        self.y_coord = DimCoord([1, 2, 3], long_name='latitude')
+        self.coords = [self.x_coord, self.y_coord]
+        self.upper = 100.
+        self.lower = 0.
+
+    def test_nobounds(self):
+        """test creating cube with neither upper nor lower threshold set"""
+        result = _make_mask_cube(self.mask, self.key, self.coords)
+        self.assertEqual(result.coord('longitude'), self.x_coord)
+        self.assertEqual(result.coord('latitude'), self.y_coord)
+        self.assertArrayEqual(result.data, self.mask)
+        self.assertEqual(result.attributes['Topographical Type'],
+                         self.key.title())
+
+    def test_upperbound(self):
+        """test creating cube with upper threshold only set"""
+        result = _make_mask_cube(self.mask, self.key, self.coords,
+                                 upper_threshold=self.upper)
+        self.assertEqual(result.coords('topographic_bound_upper')[0].points,
+                         self.upper)
+
+    def test_bothbounds(self):
+        """test creating cube with both thresholds set"""
+        result = _make_mask_cube(self.mask, self.key, self.coords,
+                                 upper_threshold=self.upper,
+                                 lower_threshold=self.lower)
+        self.assertEqual(result.coords('topographic_bound_upper')[0].points,
+                         self.upper)
+        self.assertEqual(result.coords('topographic_bound_lower')[0].points,
+                         self.lower)
+
 
 
 class Test_find_standard_ancil(IrisTest):
@@ -100,6 +142,9 @@ class Test_find_standard_ancil(IrisTest):
         msg = "That file doesn't exist"
         with self.assertRaisesRegexp(IOError, msg):
             find_standard_ancil(self.stage, msg)
+
+if __name__ == "__main__":
+    unittest.main()
 
 if __name__ == "__main__":
     unittest.main()
