+++ lib/improver/tests/utilities/test_BasicThreshold.py
@@ -1,4 +1,4 @@
-# -*- coding: utf-8 -*-
+# -*- coding: iso-8859-1 -*-
 # -----------------------------------------------------------------------------
 # (C) British Crown Copyright 2017 Met Office.
 # All rights reserved.
@@ -42,7 +42,7 @@ import numpy as np
 from improver.utilities.threshold import BasicThreshold as Threshold
 
 
-class TestThreshold(IrisTest):
+class Test_process(IrisTest):
 
     """Test the thresholding plugin."""
