+++ lib/improver/tests/utilities/test_cube_manipulation.py
@@ -1,4 +1,4 @@
-# -*- coding: utf-8 -*-
+# -*- coding: iso-8859-1 -*-
 # -----------------------------------------------------------------------------
 # (C) British Crown Copyright 2017 Met Office.
 # All rights reserved.
@@ -67,178 +67,6 @@ def _check_coord_type(cube, coord):
         coord_scalar = False
         coord_aux = True
     return coord_scalar, coord_aux
-
-
-class Test_concatenate_cubes(IrisTest):
-
-    """Test the concatenate_cubes utility."""
-
-    def setUp(self):
-        """Use temperature cube to test with."""
-        self.cube = set_up_temperature_cube()
-
-    def test_basic(self):
-        """Test that the utility returns an iris.cube.Cube."""
-        result = concatenate_cubes(self.cube)
-        self.assertIsInstance(result, Cube)
-
-    def test_identical_cubes(self):
-        """
-        Test that the utility returns the expected error message,
-        if an attempt is made to concatenate identical cubes.
-        """
-        cubes = iris.cube.CubeList([self.cube, self.cube])
-        msg = "An unexpected problem prevented concatenation."
-        with self.assertRaisesRegexp(ConcatenateError, msg):
-            concatenate_cubes(cubes)
-
-    def test_cubelist_type_and_data(self):
-        """
-        Test that the utility returns an iris.cube.Cube with the expected
-        resulting data, if a CubeList containing non-identical cubes
-        (different values for the time coordinate) is passed in as the input.
-        """
-        cube1 = self.cube.copy()
-        cube2 = self.cube.copy()
-
-        cube3 = self.cube.copy()
-        cube3.transpose([1, 0, 2, 3])
-        expected_result = np.vstack([cube3.data, cube3.data])
-
-        cube2.coord("time").points = 402195.0
-
-        cubelist = iris.cube.CubeList([cube1, cube2])
-
-        result = concatenate_cubes(cubelist)
-        self.assertIsInstance(result, Cube)
-        self.assertArrayAlmostEqual(expected_result, result.data)
-
-    def test_cubelist_different_number_of_realizations(self):
-        """
-        Test that the utility returns an iris.cube.Cube with the expected
-        realizations, if a CubeList containing cubes with different numbers
-        of realizations are passed in as the input.
-        """
-        cube1 = self.cube.copy()
-
-        cube3 = iris.cube.CubeList([])
-        for cube in cube1.slices_over("realization"):
-            if cube.coord("realization").points == 0:
-                cube2 = cube
-            elif cube.coord("realization").points in [1, 2]:
-                cube3.append(cube)
-        cube3 = cube3.merge_cube()
-
-        cubelist = iris.cube.CubeList([cube2, cube3])
-
-        result = concatenate_cubes(cubelist)
-        self.assertIsInstance(result, Cube)
-        self.assertArrayAlmostEqual(
-            result.coord("realization").points, [0, 1, 2])
-
-    def test_cubelist_different_number_of_realizations_time(self):
-        """
-        Test that the utility returns the expected error message, if a
-        CubeList containing cubes with different numbers of realizations are
-        passed in as the input, and the slicing done, in order to help the
-        concatenation is only done over time.
-        """
-        cube1 = self.cube.copy()
-
-        cube3 = iris.cube.CubeList([])
-        for cube in cube1.slices_over("realization"):
-            if cube.coord("realization").points == 0:
-                cube2 = cube
-            elif cube.coord("realization").points in [1, 2]:
-                cube3.append(cube)
-        cube3 = cube3.merge_cube()
-
-        cubelist = iris.cube.CubeList([cube2, cube3])
-
-        msg = "failed to concatenate into a single cube"
-        with self.assertRaisesRegexp(ConcatenateError, msg):
-            concatenate_cubes(cubelist, coords_to_slice_over=["time"])
-
-    def test_cubelist_slice_over_time_only(self):
-        """
-        Test that the utility returns an iris.cube.Cube with the expected
-        time coordinate, if a CubeList containing cubes with different
-        timesteps is passed in as the input.
-        """
-        cube1 = self.cube.copy()
-        cube2 = self.cube.copy()
-
-        cube2.coord("time").points = 402195.0
-
-        cubelist = iris.cube.CubeList([cube1, cube2])
-
-        result = concatenate_cubes(
-            cubelist, coords_to_slice_over=["time"])
-        self.assertIsInstance(result, Cube)
-        self.assertArrayAlmostEqual(
-            result.coord("time").points, [402192.5, 402195.0])
-
-    def test_cubelist_slice_over_realization_only(self):
-        """
-        Test that the utility returns an iris.cube.Cube with the expected
-        realization coordinate, if a CubeList containing cubes with different
-        realizations is passed in as the input.
-        """
-        cube1 = self.cube.copy()
-        cube2 = self.cube.copy()
-
-        cube2.coord("time").points = 402195.0
-
-        cubelist = iris.cube.CubeList([cube1, cube2])
-
-        result = concatenate_cubes(
-            cubelist, coords_to_slice_over=["realization"])
-        self.assertIsInstance(result, Cube)
-        self.assertArrayAlmostEqual(
-            result.coord("realization").points, [0, 1, 2])
-
-    def test_cubelist_with_forecast_reference_time_only(self):
-        """
-        Test that the utility returns an iris.cube.Cube with the expected
-        resulting data, if a CubeList containing cubes with different
-        forecast_reference_time coordinates is passed in as the input.
-        This makes sure that the forecast_reference_time from the input cubes
-        is maintained within the output cube, after concatenation.
-        """
-        cube1 = self.cube.copy()
-        cube2 = self.cube.copy()
-        cube2.coord("time").points = 402195.5
-        time_origin = "hours since 1970-01-01 00:00:00"
-        calendar = "gregorian"
-        tunit = Unit(time_origin, calendar)
-        cube1.add_aux_coord(
-            DimCoord([402192.5], "forecast_reference_time", units=tunit))
-        cube2.add_aux_coord(
-            DimCoord([402195.5], "forecast_reference_time", units=tunit))
-
-        cubelist = iris.cube.CubeList([cube1, cube2])
-
-        result = concatenate_cubes(cubelist)
-        self.assertArrayAlmostEqual(
-            result.coord("forecast_reference_time").points,
-            [402192.5, 402195.5])
-
-    def test_cubelist_different_var_names(self):
-        """
-        Test that the utility returns an iris.cube.Cube, if a CubeList
-        containing non-identical cubes is passed in as the input.
-        """
-        cube1 = self.cube.copy()
-        cube2 = self.cube.copy()
-        cube2.coord("time").points = 402195.5
-
-        cube1.coord("time").var_name = "time_0"
-        cube2.coord("time").var_name = "time_1"
-
-        cubelist = iris.cube.CubeList([cube1, cube2])
-
-        result = concatenate_cubes(cubelist)
-        self.assertIsInstance(result, Cube)
 
 
 class Test__associate_any_coordinate_with_master_coordinate(IrisTest):
@@ -525,6 +353,178 @@ class Test__strip_var_names(IrisTest):
         for cube in result:
             for coord in cube.coords():
                 self.assertIsNone(coord.var_name, None)
+
+
+class Test_concatenate_cubes(IrisTest):
+
+    """Test the concatenate_cubes utility."""
+
+    def setUp(self):
+        """Use temperature cube to test with."""
+        self.cube = set_up_temperature_cube()
+
+    def test_basic(self):
+        """Test that the utility returns an iris.cube.Cube."""
+        result = concatenate_cubes(self.cube)
+        self.assertIsInstance(result, Cube)
+
+    def test_identical_cubes(self):
+        """
+        Test that the utility returns the expected error message,
+        if an attempt is made to concatenate identical cubes.
+        """
+        cubes = iris.cube.CubeList([self.cube, self.cube])
+        msg = "An unexpected problem prevented concatenation."
+        with self.assertRaisesRegexp(ConcatenateError, msg):
+            concatenate_cubes(cubes)
+
+    def test_cubelist_type_and_data(self):
+        """
+        Test that the utility returns an iris.cube.Cube with the expected
+        resulting data, if a CubeList containing non-identical cubes
+        (different values for the time coordinate) is passed in as the input.
+        """
+        cube1 = self.cube.copy()
+        cube2 = self.cube.copy()
+
+        cube3 = self.cube.copy()
+        cube3.transpose([1, 0, 2, 3])
+        expected_result = np.vstack([cube3.data, cube3.data])
+
+        cube2.coord("time").points = 402195.0
+
+        cubelist = iris.cube.CubeList([cube1, cube2])
+
+        result = concatenate_cubes(cubelist)
+        self.assertIsInstance(result, Cube)
+        self.assertArrayAlmostEqual(expected_result, result.data)
+
+    def test_cubelist_different_number_of_realizations(self):
+        """
+        Test that the utility returns an iris.cube.Cube with the expected
+        realizations, if a CubeList containing cubes with different numbers
+        of realizations are passed in as the input.
+        """
+        cube1 = self.cube.copy()
+
+        cube3 = iris.cube.CubeList([])
+        for cube in cube1.slices_over("realization"):
+            if cube.coord("realization").points == 0:
+                cube2 = cube
+            elif cube.coord("realization").points in [1, 2]:
+                cube3.append(cube)
+        cube3 = cube3.merge_cube()
+
+        cubelist = iris.cube.CubeList([cube2, cube3])
+
+        result = concatenate_cubes(cubelist)
+        self.assertIsInstance(result, Cube)
+        self.assertArrayAlmostEqual(
+            result.coord("realization").points, [0, 1, 2])
+
+    def test_cubelist_different_number_of_realizations_time(self):
+        """
+        Test that the utility returns the expected error message, if a
+        CubeList containing cubes with different numbers of realizations are
+        passed in as the input, and the slicing done, in order to help the
+        concatenation is only done over time.
+        """
+        cube1 = self.cube.copy()
+
+        cube3 = iris.cube.CubeList([])
+        for cube in cube1.slices_over("realization"):
+            if cube.coord("realization").points == 0:
+                cube2 = cube
+            elif cube.coord("realization").points in [1, 2]:
+                cube3.append(cube)
+        cube3 = cube3.merge_cube()
+
+        cubelist = iris.cube.CubeList([cube2, cube3])
+
+        msg = "failed to concatenate into a single cube"
+        with self.assertRaisesRegexp(ConcatenateError, msg):
+            concatenate_cubes(cubelist, coords_to_slice_over=["time"])
+
+    def test_cubelist_slice_over_time_only(self):
+        """
+        Test that the utility returns an iris.cube.Cube with the expected
+        time coordinate, if a CubeList containing cubes with different
+        timesteps is passed in as the input.
+        """
+        cube1 = self.cube.copy()
+        cube2 = self.cube.copy()
+
+        cube2.coord("time").points = 402195.0
+
+        cubelist = iris.cube.CubeList([cube1, cube2])
+
+        result = concatenate_cubes(
+            cubelist, coords_to_slice_over=["time"])
+        self.assertIsInstance(result, Cube)
+        self.assertArrayAlmostEqual(
+            result.coord("time").points, [402192.5, 402195.0])
+
+    def test_cubelist_slice_over_realization_only(self):
+        """
+        Test that the utility returns an iris.cube.Cube with the expected
+        realization coordinate, if a CubeList containing cubes with different
+        realizations is passed in as the input.
+        """
+        cube1 = self.cube.copy()
+        cube2 = self.cube.copy()
+
+        cube2.coord("time").points = 402195.0
+
+        cubelist = iris.cube.CubeList([cube1, cube2])
+
+        result = concatenate_cubes(
+            cubelist, coords_to_slice_over=["realization"])
+        self.assertIsInstance(result, Cube)
+        self.assertArrayAlmostEqual(
+            result.coord("realization").points, [0, 1, 2])
+
+    def test_cubelist_with_forecast_reference_time_only(self):
+        """
+        Test that the utility returns an iris.cube.Cube with the expected
+        resulting data, if a CubeList containing cubes with different
+        forecast_reference_time coordinates is passed in as the input.
+        This makes sure that the forecast_reference_time from the input cubes
+        is maintained within the output cube, after concatenation.
+        """
+        cube1 = self.cube.copy()
+        cube2 = self.cube.copy()
+        cube2.coord("time").points = 402195.5
+        time_origin = "hours since 1970-01-01 00:00:00"
+        calendar = "gregorian"
+        tunit = Unit(time_origin, calendar)
+        cube1.add_aux_coord(
+            DimCoord([402192.5], "forecast_reference_time", units=tunit))
+        cube2.add_aux_coord(
+            DimCoord([402195.5], "forecast_reference_time", units=tunit))
+
+        cubelist = iris.cube.CubeList([cube1, cube2])
+
+        result = concatenate_cubes(cubelist)
+        self.assertArrayAlmostEqual(
+            result.coord("forecast_reference_time").points,
+            [402192.5, 402195.5])
+
+    def test_cubelist_different_var_names(self):
+        """
+        Test that the utility returns an iris.cube.Cube, if a CubeList
+        containing non-identical cubes is passed in as the input.
+        """
+        cube1 = self.cube.copy()
+        cube2 = self.cube.copy()
+        cube2.coord("time").points = 402195.5
+
+        cube1.coord("time").var_name = "time_0"
+        cube2.coord("time").var_name = "time_1"
+
+        cubelist = iris.cube.CubeList([cube1, cube2])
+
+        result = concatenate_cubes(cubelist)
+        self.assertIsInstance(result, Cube)
 
 
 if __name__ == '__main__':
